fix: remove data store
