--- app/src/main/java/com/thechance/happysocialmedia/data/DataStore.kt
@@ -1,108 +1,0 @@
-package com.thechance.happysocialmedia.data
-
-import com.thechance.happysocialmedia.domain.models.User
-
-object DataStore {
-    val users = listOf(
-        User(
-            1,
-            "James Butt",
-            "https://picsum.photos/id/0/5616/3744"
-        ),
-        User(
-            2,
-            "Josephine Darakjy",
-            "https://picsum.photos/id/1/5616/3744"
-        ),
-        User(
-            3,
-            "Art Venere",
-            "https://picsum.photos/id/10/2500/1667"
-        ),
-        User(
-            4,
-            "Lenna Paprocki",
-            "https://picsum.photos/id/100/2500/1656"
-        ),
-        User(
-            5,
-            "Donette Foller",
-            "https://picsum.photos/id/1000/5626/3635"
-        ),
-        User(
-            6,
-            "Simona Morasca",
-            "https://picsum.photos/id/1001/5616/3744"
-        ),
-        User(
-            7,
-            "Mitsue Tollner",
-            "https://picsum.photos/id/1002/4312/2868"
-        ),
-        User(
-            8,
-            "Leota Dilliard",
-            "https://picsum.photos/id/1003/1181/1772"
-        ),
-        User(
-            9,
-            "Sage Wieser",
-            "https://picsum.photos/id/1004/5616/3744"
-        ),
-        User(
-            10,
-            "Kris Marrier",
-            "https://picsum.photos/id/1005/5760/3840"
-        ),
-        User(
-            11,
-            "Minna Amigon",
-            "https://picsum.photos/id/1006/3000/2000"
-        ),
-        User(
-            12,
-            "Benjamin Combs",
-            "https://picsum.photos/id/1008/5616/3744"
-        ),
-        User(
-            13,
-            "Abel Maclead",
-            "https://picsum.photos/id/101/2621/1747"
-        ),
-        User(
-            14,
-            "Kiley Caldarera",
-            "https://picsum.photos/id/1010/5184/3456"
-        ),
-        User(
-            15,
-            "Graciela Ruta",
-            "https://picsum.photos/id/1011/5472/3648"
-        ),
-        User(
-            16,
-            "Cammy Albares",
-            "https://picsum.photos/id/1012/3973/2639"
-        ),
-        User(
-            17,
-            "Mattie Poquette",
-            "https://picsum.photos/id/1013/4256/2832"
-        ),
-        User(
-            18,
-            "Meaghan Garufi",
-            "https://picsum.photos/id/1014/6016/4000"
-        ),
-        User(
-            19,
-            "Gladys Rim",
-            "https://picsum.photos/id/1015/6000/4000"
-        ),
-        User(
-            20,
-            "Yuki Whobrey",
-            "https://picsum.photos/id/1016/3844/2563"
-        )
-    )
-}
